Implemented Sessions
Now the user data is not passed through the URL.

This is important because before, anyone could change their URL to say "teacher" and would then get the privelages of a teacher, without ever having logged in as a teacher, and they could access other accounts without logging in, etc. This also means our prepare_query_string function is no longer necessary, so I deleted the two places where that function was written.