Cleaned up redirecting
There was a ton of inconsistency in how different URLs were addressed. I have made them all consistent and solved a few bugs in the process. The standard now is for a header function use "Location: filename.php", with NO SLASH at the end. For a <form action=""> tag, use the form "./filename.php" with one period, one slash, the filename, and no slash at the end.