explosion animation + other things
Added an explosion animation for solving problems.

spawnMax now starts much higher to make it more reasonable for 2nd graders.

the background image is now loaded with the canvas rather than redrawn every frame: this should make things faster.

Made sure the reset button works properly. It does not refresh the page, and the cursor is put back into the answerForm.

There is a bug with the explosions where on newly loading the page, the first problem does not have an explosion animation, but all subsequent problems do, including if you hit reset. Still have not figured this out but it's been too long so I'm pushing this stuff. The games still work fine, its just we dont see an explosion for the very first problem.