Fixed explosion bug
Explosions weren't happening on the first problem solved on a newly loaded page.

This was caused because I was trying to load and display each image within the same tick, but the images take too long to load, so they wouldn't appear the first time. However, by the second time, the loading functions have completed and the images are loaded so the next one shows.

I fixed it by loading all the explosion frames into an array called explosionFrames, which happens in the script directly, and thus when the page is loaded, and before the first problem is solved. Then in updateExplosions() I simply access the image that's part of the array.

This should be a speed improvement overall in addition to fixing the bug.