difficult -> gamemode, strikes
changed every mention of "difficulty" to "gamemode" because thats what it actually is.

got rid of the old penalty system and now you just have a certain number of strikes, or problems you can get wrong, before you lose. This better fits with the graphic of the laser bouncing off an asteroid that you get wrong, and still prevents brute-forcing the digit game.