several changes, read desc
added new background image which shows a planet at the bottom

when the player loses, the asteroids keep going, but no new ones are spawned. They explode with boss explosions (regardless of whether or not they are boss problems) upon hitting the planet.

moved the score counter onto the canvas in the top left corner

made the gameover message appear in the canvas, along with a button in the canvas which allows you to play again, which also properly focuses on the answerForm. The button also lights up when you mouse over it

added a cannon at the bottom of the screen which sits on the planet and appears to fire the laser

changed bossCountdown to simply be generated [5,10] instead of [5,15] - gamemode