optimizations to game graphics
Made the background a css element instead of part of the canvas

Split the canvas into two canvas: mid and top. topCanvas contains the score, game over message, and cannon. midCanvas contains asteroids, lasers, and explosions.

Matheroids are now rendered once when they are created, as offscreen canvases, then redrawn as images. This is faster than drawing the text directly every frame.

I may make optimizations to the explosion animation in the future, using offscreen canvases to pre-scale all the images instead of scaling them every frame. This may not be necessary