Added music and a loss sound
Made and added music. The music loops continuously starting on the page load; resetting the game without a page reload does not reset the music.

Added a sound effect which plays when you lose.

Increased strikes to 5 although this may be subject to teacher manipulation in the future

Added icons for music and sound effects so the user can choose whether to hear them