Added create teacher functionality
The create teacher button on the welcome screen now goes to new_teacher_form.php which uses create_teacher.php to write new teachers to the users file. It doesn't work completely because I am not sure of how the hashing passwords thing works as far as writing new passwords to the database.