main menu can be scrolled
teacher dashboard was getting cut off at the bottom for me since I couldn't scroll, now you can scroll.

One problem which is very minor but somehow baffling to me is that when you switch to leaderboard, teacher dashboard, or gradebook the page scrolls to the bottom, but when you switch to main or settings, the scroll position stays where it was.